+++ packages/store/tests/issues/issue-2180-dispatching-in-effect.spec.ts
@@ -7,9 +7,10 @@ import {
 } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { Action, NgxsModule, Selector, State, StateContext, Store } from '@ngxs/store';
+import { skipConsoleLogging } from '@ngxs/store/internals/testing';
 
-describe('Dispatching actions in effects (https://github.com/ngxs/store/issues/2180)', () => {
-  class Set {
+describe('State per signal', () => {
+  class SetNumber {
     static readonly type = 'Set';
 
     constructor(readonly value: number) {}
@@ -26,8 +27,8 @@ describe('Dispatching actions in effects (https://github.com/ngxs/store/issues/2
       return number;
     }
 
-    @Action(Set)
-    set(ctx: StateContext<number>, action: Set) {
+    @Action(SetNumber)
+    setNumber(ctx: StateContext<number>, action: SetNumber) {
       ctx.setState(action.value);
     }
   }
@@ -35,31 +36,35 @@ describe('Dispatching actions in effects (https://github.com/ngxs/store/issues/2
   @Component({
     template: `
       <h1>{{ value() }}</h1>
+      <h2>{{ number() }}</h2>
       <button (click)="updateValue()">Click me</button>
-    `
+    `,
+    standalone: true
   })
   class TestComponent {
     value = signal(0);
+    number = this.store.selectSignal(NumberState.getNumber);
 
-    constructor(store: Store) {
+    constructor(private store: Store) {
       effect(() => {
         const value = this.value();
-        store.dispatch(new Set(value));
+        store.dispatch(new SetNumber(value));
       });
     }
 
     updateValue() {
-      this.value.set(100);
+      this.value.update(value => value + 100);
     }
   }
 
   it('should allow dispatching actions in effects', async () => {
     // Arrange
-    TestBed.configureTestingModule({
-      declarations: [TestComponent],
-      imports: [NgxsModule.forRoot([NumberState])],
-      providers: [provideExperimentalZonelessChangeDetection()]
-    });
+    skipConsoleLogging(() =>
+      TestBed.configureTestingModule({
+        imports: [NgxsModule.forRoot([NumberState]), TestComponent],
+        providers: [provideExperimentalZonelessChangeDetection()]
+      })
+    );
 
     const fixture = TestBed.createComponent(TestComponent);
     fixture.detectChanges();
@@ -68,8 +73,11 @@ describe('Dispatching actions in effects (https://github.com/ngxs/store/issues/2
     // Act
     document.querySelector('button')!.click();
     await fixture.whenStable();
+    document.querySelector('button')!.click();
+    await fixture.whenStable();
 
     // Assert
-    expect(document.querySelector('h1')!.innerHTML).toEqual('100');
+    expect(document.querySelector('h1')!.innerHTML).toEqual('200');
+    expect(document.querySelector('h2')!.innerHTML).toEqual('200');
   });
 });